py/requirements.txt
langchain
langchain-openai
langchain-community
langgraph
openai>=1.10.0
pillow
requests
PyYAML>=6.0
python-dotenv 
flask
fastapi
uvicorn
redis>=4.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
